Fixed flaws in current game state.
The materials were not on their respective objects due to the project updating to Unity 6. I also had to fix an issue with the angle in which the ball took after being thrown, it now goes forward at the angle in which it is tossed. If the player is able to get the ball into the hoop it adds to the count and reflects that via a UI counter text.