Updated the Player script
Added response to an interact action, currently nothing can be collected but the basis is made. The player detects what is near them that can be collected (it currently detects the borders of the game scene) and when the interact button is pressed they will get a response via debug.log.